Parsing, Headless Chrome
Created two parsing methods inside of OptimizeResume, each attempts to divide the resume into sections and extract key information.

I am planning on implementing some kind of web scraping to perform live research based on the job posting, and tailor optimizations accordingly, so I updated the build to install a headless chrome browser for linux and add it to the path.